Now you can't delete a category if it has trackers in it

+++ Tracker/Domain/TrackerCategoryStore.swift
@@ -17,6 +17,7 @@ enum TrackerCategoryStoreError: Error {
     case fetchCategoriesFailed
     case getContextError
     case categoryDeleteError
+    case categoryEditError
     case otherError
 }
 
@@ -63,7 +64,7 @@ final class TrackerCategoryStore: NSObject {
     private var categories = [TrackerCategory]()
     
     // MARK: - Public Methods
-    public func fetchCategories() throws -> [TrackerCategory] {
+    func fetchCategories() throws -> [TrackerCategory] {
         guard let objects = fetchedResultsController?.fetchedObjects else {
             throw TrackerCategoryStoreError.fetchCategoriesFailed
         }
@@ -72,7 +73,7 @@ final class TrackerCategoryStore: NSObject {
         return categories
     }
     
-    public func addCategory(title: String) throws {
+    func addCategory(title: String) throws {
         guard let context else { throw TrackerCategoryStoreError.getContextError }
         //check unique
         let fetchRequest = TrackerCategoryCoreData.fetchRequest()
@@ -92,18 +93,38 @@ final class TrackerCategoryStore: NSObject {
         appDelegate?.saveContext()
     }
     
-    public func deleteCategory(title: String) throws {
+    func editCategory(title: String, for existingTitle: String) throws {
+        guard let context = context else { throw TrackerCategoryStoreError.getContextError }
+        
+        let fetchRequest = TrackerCategoryCoreData.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "%K == %@", #keyPath(TrackerCategoryCoreData.title), existingTitle)
+        
+        let findCategory = try? context.fetch(fetchRequest)
+        guard let categoryToEdit = findCategory?.first else {
+            throw TrackerCategoryStoreError.categoryEditError
+        }
+        
+        categoryToEdit.title = title
+        
+        appDelegate?.saveContext()
+    }
+    
+    
+    func deleteCategory(title: String) throws {
         guard let context else { throw TrackerCategoryStoreError.getContextError }
         
         let fetchRequest = TrackerCategoryCoreData.fetchRequest()
         fetchRequest.predicate = NSPredicate(format: "%K == %@", #keyPath(TrackerCategoryCoreData.title), title)
         
         let findCategory = try? context.fetch(fetchRequest)
-
+        
         guard let findCategory else { throw TrackerCategoryStoreError.categoryDeleteError }
         
         if (findCategory.first) != nil {
             guard let category = findCategory.first else { throw TrackerCategoryStoreError.categoryDeleteError }
+            if category.trackers?.count ?? 0 > 0 {
+                throw TrackerCategoryStoreError.otherError
+            }
             context.delete(category)
         } else {
             throw TrackerCategoryStoreError.categoryDeleteError
